Create stack-implementation-c
The purpose of this project is to understand the usage of .h and .c files and the concept of prototype in C.

I implemented a stack by using an array in this project. My program should be able to use the stack to hold a list of integers and print out the list in the original order and the reverse order.